+++ crdesigner/ui/gui/start_gui.py
@@ -1,7 +1,7 @@
 import sys
 from pathlib import Path
 
-from PyQt5.QtWidgets import *
+from PyQt6.QtWidgets import *
 from crdesigner.ui.gui.controller.mwindow_controller import MWindowController
 from crdesigner.ui.gui.utilities.file_actions import open_path
 
@@ -17,7 +17,7 @@ def start_gui(input_file: str = None):
     if input_file:
         open_path(mwindow=w, path=input_file)
     w.mwindow_ui.showMaximized()
-    sys.exit(app.exec_())
+    sys.exit(app.exec())
 
 
 if __name__ == '__main__':